+++ src/components/BookSearch/BookSearch.jsx
@@ -1,13 +1,13 @@
 import BookSearchForm from '../BookSearchForm/BookSearchForm';
 
 
-const BookSearch = ({fetchData, selectedSearchType,setSelectedSearchType,handleSearch, categories, handleCategoryChanges,setStartIndex, query, setQuery,setCurrentPage}) => {
+const BookSearch = ({fetchData, selectedSearchType,setSelectedSearchType, categories, handleCategoryChanges,setStartIndex, query, setQuery,setCurrentPage}) => {
   
   return (
     <div style={{ border: '1px solid black' }}>
       <BookSearchForm
         
-        handleSearch={handleSearch}
+        
         categories={categories}
         handleCategoryChanges={handleCategoryChanges}
         setStartIndex={setStartIndex}
