+++ src/components/Footer/Footer.jsx
@@ -6,7 +6,7 @@ const Footer = () => {
     <div className='footer'>
     <Typography variant="body2" color="white" align="center" component="footer">
       {'Copyright © '}
-        Book Snap{' '}
+      BookSnap{''} by Annabelle Jenny Phuong Yonghee  {''}
         {new Date().getFullYear()}
       {''}
     </Typography>
